% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_def.R
\name{add_coefs_to_formula}
\alias{add_coefs_to_formula}
\title{Add (optionally randomly-generated) coefficients to a formula in a definition data.table.}
\usage{
add_coefs_to_formula(chr_formula, coefs = NULL, u_coef = NULL)
}
\arguments{
\item{chr_formula}{Character vector, of the form \code{'x1+x2+x3'}}

\item{coefs}{Numeric vector, coefficients, length equal to number of terms (including intercept)}

\item{u_coef}{Numeric length 1, coefficient for the variable \code{u}, which we fix to be constant over time to ensure parallel trends}
}
\value{
Another character vector with coefficients added to formulas, of the form \code{'a + b*x1 + c*x2 + d*x3'}
}
\description{
Add (optionally randomly-generated) coefficients to a formula in a definition data.table.
}
\examples{
add_coefs_to_formula('x1+x2+x3')
}
