% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_def.R
\name{default_link}
\alias{default_link}
\title{Get default link for a distribution}
\usage{
default_link(dist)
}
\arguments{
\item{dist}{chr. Currently 'normal', 'binomial', or 'binary'.}
}
\value{
Character vector containing link names.
}
\description{
Get default link for a distribution
}
\examples{
default_link(c('normal','binary'))
}
