% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_panel_plot.R
\name{did_panel_plot}
\alias{did_panel_plot}
\title{generate a DiD panel plot}
\usage{
did_panel_plot(data, unit, year, group, treat)
}
\arguments{
\item{data}{A df of DiD panel data with unit, group, year and treatment status indicators}

\item{unit}{character vector specifying the name of the unit variable}

\item{year}{character vector specifying the name of the year variable}

\item{group}{character vector specifying the name of the group variable}

\item{treat}{character vector specifying the name of the treatment variable}
}
\value{
a ggplot object
}
\description{
generate a DiD panel plot
}
