% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_plot.R
\name{did_plot

did_plot generates a plot showing group outcome means over time}
\alias{did_plot

did_plot generates a plot showing group outcome means over time}
\alias{did_plot}
\title{generate a DiD plot}
\usage{
did_plot(data, year, y, group)
}
\arguments{
\item{data}{A df of DiD panel data with year, outcome and group indicators}

\item{year}{character vector specifying the name of the year variable}

\item{y}{character vector specifying the name of the outcome variable}

\item{group}{character vector specifying the name of the group variable}
}
\value{
a ggplot object
}
\description{
generate a DiD plot
}
