% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_def_extras.R
\name{intervene}
\alias{intervene}
\title{Intervene on a data definition}
\usage{
intervene(def, varnames, values)
}
\arguments{
\item{def}{Definition data.table to be modified}

\item{varnames}{Character vector. Names of variables whose value the intervention will modify.}

\item{values}{Numeric vector, same length as \code{varnames}. Values those variables will take under the intervention.}
}
\value{
A data.table that is an updated data definitions table.
}
\description{
Intervene on a data definition
}
\examples{
def = pt_def(formulas = list(w1 = 'w1{t-1} + a{t-1}',
                             a  = 'w1{t} + u',
                             y  = 'w1{t} + a{t} + u'),
             dists = c('normal','binary','normal'),
             variances=c(1,1,1), periods = 3, u_coef = 1)

def0 = intervene(def, varnames = c('a1','a2'), values = c(0,0))

}
