% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_def.R
\name{pt_def}
\alias{pt_def}
\title{Create a definitions table for a \code{simstudy} simulation design under parallel trends}
\usage{
pt_def(
  formulas,
  dists,
  links = default_link(dists),
  variances = rep(1, length(formulas)),
  periods,
  u_coef = 1
)
}
\arguments{
\item{formulas}{named list.}

\item{dists}{character vector.}

\item{links}{character vector.}

\item{variances}{numeric vector.}

\item{periods}{integer.}

\item{u_coef}{numeric vector of length 1 or \code{length(formulas)}.}
}
\value{
A data.table that is a data definitions table for simstudy.
}
\description{
Create a definitions table for a \code{simstudy} simulation design under parallel trends
}
\examples{
formulas = list(
  w1 = 'w1{t-1} + a{t-1}', #note no intercepts - these will be added automatically
  w2 = 'w1{t}   + w2{t-1} + a{t-1}',
  w3 = 'w1{t}   + w2{t}   + w3{t-1} + a{t-1}',
  a  = 'w1{t}   + cos(w2{t})   + I(w3{t}^2)  + u',
  y  = 'sin(w1{t})   + w2{t}*w3{t}   + a{t} + u'
)
dists =  c(
 w1 = 'normal',
 w2 = 'normal',
 w3 = 'normal',
 a  = 'binary',
 y  = 'normal'
)
variances = c(
  w1 = .1,
  w2 = .1,
  w3=  .1,
  a = 1, #this has no effect
  y = 0.1
)
pt_def(formulas, dists, links=default_link(dists),
       variances, periods = 3, u_coef = 1)


}
