% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app

run_app is a wrapper for the ui,server and call elements of the shinyApp}
\alias{run_app

run_app is a wrapper for the ui,server and call elements of the shinyApp}
\alias{run_app}
\title{shinyApp for staggered DiD simulation}
\usage{
run_app()
}
\value{
runs shinyApp
}
\description{
shinyApp for staggered DiD simulation
}
